Rename 'TextBuffer' to 'TextBufferState' to prepare for wrap in MVar
The 'TextBufferState' (previously 'TextBuffer') is an ordinary
immutable data type with bookeeping information pertaining to, and
references to, mutable vectors. So when an update to the mutable
vector occurs the 'TextBufferState' needs to be updated. So it is
important to keep updated bookeeping information in an MVar that way a
user of this module don't have to always remember to keep track of the
latest version of the 'TextBufferState'.