Add accounting fields for line break characters.
A line break can be formed from 1 or 2 characters, for example "\n" or
"\n\r". However most text editors treat line breaks as single
characters no matter how many characters are actually used in the line
break instance. In order for cursor motion functions to behave in this
way, there needs to be a way to skip past the line break in a single
action, therefore the number of characters to skip is recorded with
each line.