Bug fix WIP: Solved slice range exception, trim on first&last line.
Bugs still exist, debugging 'trace' statements are still written into
the code in this commit.

The 'slice' function was throwing an exception, I was passing an
absolute index rather than a slice length. This is fixed.

The 'trim' function is a local function that copies part of a TextLine
to the target 'TextView'. A text position in the middle of the line at
the start of the selection should cut from the middle of the line to
the end of the line, a text position in the middle of the line at the
end of the selection should cut from the start of the line to the
middle of the line. The opposite was happening: the start of the
selection was cutting from zero to the middle, the end selection was
cutting from the middle to the end. This has been fixed.