Re-renamed 'resumeEditLine' back to 'editLine', it was a mistake.
It was a big mistake to make 'editLine' force an 'beginInsertMode'
state change, it has been causing so much confusion. In fact, I am
starting to question whether 'beginInsertMode' and 'endInsertMode'
should even exist in this API. Perhaps 'endInsertMode' should be
changed to 'writeLineToCursor', and have it called automatically only
when the cursor moves and when a change has been made to the line
editor since it was last moved.