Bug fix: '\NUL' characters no longer appended to 'TextLine's
The 'insertString' function was calling 'clearCurrentLine', which does
is a user-facing function that not delete the line break
characters. The problem with this is that 'insertString' does not push
line breaking characters after the cursor, only before, and it also
assumes 'clearCurrentLine' deletes the entire line, including the line
break characters. This resulted in 'clearCurrentLine' resetting the
length of the region after the cursor to the number of line breaking
characters as though there were line breaking characters pushed after
the cursor. Since there were never any line breaking characters pushed
after the cursor but the 'clearCurrentLine' function was always
setting the size of the after-cursor region to a value of 1, this
resulted in a single null character always being appended to the
'TextLine' when 'copyLineEditor'' was called.