Created 'pushElemVec' function, replacing 'unsafeInsertString'.
This function makes use of the 'GVec.copy' API, so is probably going
to be much more efficient than using the for-loop method of
index-by-index copying that the old 'unsafeInsertString' function
used. This update also eliminated the last function that used the
'relativeToLine', 'relativeToChar', and 'growVec' functions. All have
been removed.