Bug fix: cannot use 'bracket' for 'listFSNodes'.
Error handling is now done with a catch and re-throw after closing the
directory stream. Handling open files with pipes which keep the
handles open until the last item in the pipe is going to lead to some
pretty big problems down the line, it could make this whole experiment
useless if I don't figure out a way to deal with Piping elements out
of resources like open files or directories.