Pipe is NOT a Monad, it is an Applicative Functor.
It was a mistake to try and instantiate 'Pipe' into the various Monad
type classes (Monad, MonadPlus, MonadError, and MonadIO). I realized
this mistake when I realized that an expression
like (msum[pipe[1,2,3],empty,pipe[4,5,6]]) was equivalent
to (pipe[1,2,3,4,5,6]) but was not equivalent
to (join(pipe[pipe[1,2,3],empty,pipe[4,5,6]])).
Instead (join(pipe[pipe[1,2,3],empty,pipe[4,5,6]])) is equivalent
to (pipe[1,2,3]) because of how 'join' uses monadic bind, and how
monadic bind halts when it binds to an 'empty' function. I could not
think of how to extend the constructors of (Pipe m a) in such a way
that would have made the behavior of 'join' and 'msum' equivalent, but
not in such a way that 'Pipe' was identical to a 'ListT' monad
transformer. I finally realized that there is no such way, because the
operational semantics of 'Pipe' that I need are different from the
'ListT' monad transformer, and that my operational semantics are
defined such that there is no way to modifiy the 'Pipe' constructor
lists such that 'Pipe' is a Monad. It is good enough for 'Pipe' to be
in the Applicative, Alternative, and Functor typeclasses. For monadic
behavior controlled by 'Pipe's, 'EngineT' should be used instead.