Renamed 'PipeLink' to 'CPS'.
I thought 'PipeLink' might be easier to remember, but I worry that the
name might be confusing -- users may have trouble remembering when to
use a 'Pipe' and when to use a 'PipeLink'. So I decided to rename it
to 'CPS', which is perhaps more essoteric, but also more concise.