Bugs fixed: cursorIndex, getRelToAbs, getSlice, getSliceChk
I have been horribly inconsistent in computing the index value of the
cursor. I must remember that the index value is always equal to the
number of elements before the cursor subtracted by one, and that the
number of elements before the cursor points to a space between
the top-most low element and the bottom-most undefined element.