WIP: Created read-only 'ViewChars' and 'ViewLines' functions
My hope is to instantiate these functions into 'MonadEditVec' such
that I can make use of the existing (and already tested and correctly
working) generic 'MonadEditVec' functions to slice and splice
immutable vectors just as well as I can do with mutable vectors.

Unfortunately, the 'newVector' function, used to create copies of
vectors, must be mutable, so the 'newVector' function may need to be
split off into a separate class, maybe call the class 'MonadCopyVec'
so that 'copyRegion' can use declare in it's type constraints context
that it make use of both classes.