Bug fix: 'insertString' must not call 'clearCurrentLine'
The 'clearCurrentLine' function is a public-facing API that does not
delete the contents of the line editor completely: it does not clear
the line break because that would require the next line be copied into
the line editor, because you can't have two consecutive lines not
broken by a line break. However the 'insertString' function does
actually need the entire line editor to be cleared. The bug manifested
itself like this:

The 'insertString' function never inserts characters after the cursor,
the number of characters after the cursor must always be zero. After
freezing the first line of input, the 'clearCurrentLine' function is
called, setting the number of characters after the cursor to be the
the number of line-breaking characters, which is non-zero on
multi-line input. On the next line of input, the number of characters
after the cursor has not been reset to zero, even though no characters
have been inserted after the cursor, so every line of text after the
first line passed to 'insertString' has extraneous null characters
equal to the number of line-breaking characters inserted into the line
editor buffer.