gotoCurosr function was redundant, renamed getCursor to getPosition
Chances are I have renamed these functions before because I couldn't
remember that they were named 'getPosition' and 'gotoPosition' and
thought 'getCursor' and 'gotoCursor' were more intuitive. But I am
going to have to stick with 'getPosition' and 'gotoPosition' because
we aren't really getting or going-to a cursor object, we are getting
or going to a location given by a position value.