Bug fix: do not flush line editor before creating a 'TextView'
I was using copyLineEditorText' to flush the LineEditor, which does
not check if the line editor is up-to-date with the latest cursor
movements. It may be a good idea to flush the line editor before
creating a vew (or it might be a better idea to let users of this
library decide), but if it does it mus ensure that 'insertLine' and
'moveCursor' have properly re-populated the 'LineEditor' and check
that the 'LineEditor' has been updated before actually copying the
'LineEditor' content to the current cursor position.