Reformulated test setup such that only failed tests are run
This required following instructions at url
<https://hspec.github.io/rerun.html>. I also needed to change the
'textViewTests' so that the test buffer is initialized with the text
grid during the test initializatinon phase using 'runIO'.

Prior to this commit, initializing the test buffer and filling it with
a test grid pattern was itself a test (declared with the 'it'
function, rather than the 'runIO' function) which would either pass or
fail. The problem is if only the tests that failed are re-run, as it
is now configured to do, the buffer is initialized but not filled with
a test grid pattern because that test passed and isn't re-run. This
causes other related tests to fail (even when they are correct) when
tests are re-run, as these test were written assuming the grid pattern
exists.

Since initializing the buffer and filling it with the test grid
pattern makes use of 'insertString', I re-ordered the tests so that
'insertString' is tested first by the 'textEditorTests' function, and
that 'insertString' is not tested for correctness but for whether or
not it will fail during the initialization phase. Once there is
confidence that that the 'insertString' function will not throw an
exception, the 'textViewTests' which rely on 'insertString' are run in
a separate test case.