Subgroups textView tests, each starting with a call to gotoPosition
I am not sure the tests before were correct, the call to
'gotoPosition' was in it's own test case separate from the 'vi' tests
(in the 'textViewTests' function), meaning when Hspec re-runs the test
suite only for failed tests, the 'gotoPosition' test cases were not
being called.

I created a sub-group of tests (each sub-group begins with a
'describe' function), and the sub-test begins by moving the cursor --
not in the first group, the cursor is left as it is immediately after
initialization. Each sub-group moves the cursor then creates a series
of 'textView's which hopefully test the edge-cases of creating views
around a cursor that splits the buffer.