Bug fix: use cursorIndex not (getElemCount Before), tests still fail
Also use 'diffAbsolute' to do relative index arithmentic. This bug fix
only corrects some cursor motion functions.

The reason the tests still fail is that I just realized I have been
operating on the assumption that the line cursor, which is a line
editor, always has a copy of the 'TextLine' that is stored at the
'cursorIndex', which is not true. In fact, I had designed this editor
such that loading the 'TextLine' under the cursor does NOT happen
automatically so as to allow programmers to decide when to copy
characters, and then I go and write my code on the assumption that the
text line cursor automatically loads the 'TextLine' under the cursor.

The solution will be to allow the line editor cursor to move around
even if the line editor buffer is empty. Care must be taken to ensure
that moving the cursor on a stale line editor does not result in the
line editor copying uninitialized characters into the text buffer.

I need to finally write a 'flushTextCursor' function.