Improved 'modifyColumn' to be less buggy, but tests still fail.
The behavior is vastly improved from before, when I was using the
'withCurrentLine' function within the 'modifyColumn' function which
was throwing a soft out-of-bounds exception. The "bug" was caused by
the fact that the buffer was empty for the test cases and
'withCurrentLine' was trying to access the line under the cursor, but
the line under the cursor that doesn't exist. (Because no lines exist
in an empty buffer!)

I decided that this behavior wasn't a bug, that the 'withCurrentLine'
function should throw an exception when used on an empty buffer. So I
decided to change the 'modifyColumn' function instead, this time to
check if the line editor is "unclean" and, if so, operate on the
'LineEditor' directly rather than calling 'withCurentLine'. This
solves the problem of the exception being thrown because an empty
'TextBuffer' is always (or should always be) "unclean". If there are
lines in the 'TextBuffer' and the line editor is "clean", then
'withCurrentLine' is used to check the bounds of the cursor movement
as it was doing before.

The tests are still mostly failing, but it now seems to be due to a
fencepost error of some kind. At least these test functions are
returning results that are closer to the correct answers now.