Swapped content of 'lineEditorPreTests' and 'lineEditorTests'
The 'lineEditorTests', which was 'lineEditorPreTests' and performs the
same tests as before (only the name has changed) is now a single unit
test. Before this commit, this test function consisted of three unit
tests, each dependent on the previous test to be correct in order to
pass itself.

Interdependencies between unit tests does not fit well with the Hspec
ethos. Interdepdent tests should be all placed into a single unit
test, and in fact there is no easy way to decalre such dependencies
with the Hspec API. Furthermore, enabling the '--rerun' option to
Hspec may disable tests that passed but which are required to run if
other dependent test cases have not passed yet, resulting in depdent
tests always failing even when they are correct.