Bug fix: only use internal vector editor ops to do primitive ops
Especially, do not use (currentLineNumber >>= getLineIndex) to get the
current line when (getElem Before) will do the job.

The fact that (currentLineNumber >>= getLineIndex) doesn't work is a
bug, however. I am going to have to fine-tune the user-facing
primitives to behave in more intuitive ways, you can't just wrap the
vector editor ops with bounds-checking code, this is too prone to
fencepost errors.