Bug fix: in 'textView', checking minBound and maxBound values.
This user-facing function can receive 'TextLocation' values which have
'LineIndex' or 'CharIndex' components that may have be equal
'minBound' or 'maxBound' values. As it was written, it did check if
the given 'CharIndex' values were 'minBound' or 'maxBound' but it was
not checking the given 'LineIndex' values. This patch fixes that.