Bug fix: DFS algorithm closes directories during the correct step.
Prior to this patch, directory streams would be opened and remain open
even after the final file element from the directory was read, leading
to the exhaustion of the scarce file descriptor resources allocated to
the process by the operating system when traversing a directory tree
with a large number of directories.

I believe the problem may have been caused by the 'ContT' continuation
monad transformer, and how I unwrapped the transformer to evaluate the
directory traversal loop. The way it was written, the 'closeDirStream'
function was not being evaluated until after the entire traversal
completed, but I am not certain why this happens. I moved the point
'closeDirStream' is evaluated within the algorithm to be evaluated the
moment immediately after the 'readDirStream' function evaluates to a
null string.