WIP: made progress fixing the 'deleteCharsWrap' bug, still not fixed
I have fairly thoroughly analyzed the problem and written comments in
the code explaining everything I've learned so far.

There is still a bug: it appears that the 0th element of the array is
being over-written by a blank line, but after evaluating
'debugPrintBuffer' before and after 'deleteCharsWrap', the buffer is
in exactly the correct state. It seems the problem is probably with
the 'textView' function, even though I thought I had thoroughly tested
this function. I really need to write some QuickCheck tests.