Bug fix: 'textView' correctly handles minBound and maxBound ranges
Also added some pre-tests checking for 'minBound' and 'maxBound'
conditions. Also, instantiated 'TextLocation' into the Bounded
typeclass.

Be very careful when evaluating 'charToIndex' on a 'TextLocation'
value, especially one that has not yet been bounds checked. If either
part of the 'TextLocation' is 'maxBound' or 'minBound', the
'charToIndex' conversion could cause the value to overflow.