WIP: 'deleteCharsWrap' shall call 'deleteAllChars' on min/max bound
Deleting to an extreme character count like 'minBound' or 'maxBound'
should be a simpler algorithm and should be implemented separately,
perhaps called 'deleteAllChars'. The 'deleteCharsWrap' function should
check for this condition and call 'deleteAllChars'. This will also
allow the remainder of 'deleteCharsWrap' to compute character counts
without using 'safeAbs' (which is 'abs' but always checks for
'minBound').