Redefine 'textView' in terms of 'validateLocation'.
Also, the code for checking whether an (Absolute CharIndex) is
in-bounds with respect to a Textine value has been split out of the
'testLocation' function and into it's own pure 'testCharIndex'
function. The 'textCharIndex' function has been made public because it
will almost certainly be useful to end users for checking indicies
against 'TextLine's.