Bug fix: stateful (from <- pure (min from to)) is a bad idea
I was overwriting the 'to' location with the 'from' locaiton, only in
the situation when the 'from' was greater than 'to'. The whole point
of using Haskell is to avoid stateful mistakes like this, but if you
don't take advantage of purity as often as you can, you're bound to
shoot yourself in the foot like I just did here.