Make semantics of 'streamLook' and 'streamStep' more orthogonal
Where before, 'streamLook' did the same thing as 'streamStep' without
advancing the cursor, now 'streamStep' only advances the cursor and
'streamLook' must be called explicitly in order to obtain the cursor
before advancing with 'streamStep'. This was necessary when I realized
how the `parsers` library API broke down parser functionality, and how
I needed to make explicit when the character under the parser's cursor
is read, and when the cursor is moved.