Cursor index shall ALWAYS equal number of elements before the gap.
I am going to have endless fencpost errors after making this
change. But after writing my command line program around this
TextEditor, I have realized that it is essential that the semantics of
these API functions treat the cursor index value as being equal to the
number of elements before the gap in the gap buffer.