Realized a major error in my thinking about this data structure
The number of elements and the index of the cursor are not always
equal because the index can be incremented to an undefined element. I
had considered this before, and was keeping the cursor index and the
element count separate values, but I got confused because it seemed
the values were redundant most of the time, so I removed the cursor
index.

Now I understand the problem, and am solving the problem in a less
confusing way: there will be a boolean flag set to indicate whether
the element under the cursor is defined. This flag will be checked
when counting elements and adjust the count by 1 when necessary.

I will have to review most of my code and make sure the assumptions I
have making before are effected by this change.

squash