Create ViewText function type, consistently name ViewLine functions
In creating the ViewText type, I am trying to rename some of the
functions that operate on the ViewLine type so that there is a
consistent naming scheme between the two types. Ultimately, I'd like
to make as many of the ViewText type functions that have similar types
and semantics to many of the EditText type functions into a single
typeclass.

These changes have prompted me to review the type contexts of the
various functions in the of "Hakshell.GapBuffer" module, make them
cleaner and more consistent.