Get rid of "FoldMap" module, redefine RewriteLines function type
The function for rewriting lines is no longer a wrapper around
'FoldMapLines', it has it's own unique monad transformer
within. Rewrote 'deleteCharsWrap' using more elegant algorithm using
the new 'RewriteLines' function.

Changed fundamental nature of 'pushLine': it checks if the 'TextLine'
being pushed contains a line break, and if not, append or prepend the
'TextLine' to the current 'LineEditor' depending on whether the
'RelativeToCursor' is 'Before' or 'After'.